refactor/begin implementation of date range
