+++ internal/database/podcastRepository.go
@@ -1,0 +1,37 @@
+package database
+
+import (
+	"github.com/pkg/errors"
+	"gorm.io/gorm"
+	"ikoyhn/podcast-sponsorblock/internal/models"
+)
+
+func PodcastExists(podcastId string) (bool, error) {
+	var episode models.Podcast
+	err := db.Where("id = ?", podcastId).First(&episode).Error
+	if err != nil {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return false, nil
+		}
+		return false, err
+	}
+	return true, nil
+}
+
+func GetPodcast(id string) *models.Podcast {
+	var podcastDb models.Podcast
+	err := db.Where("id = ?", id).Find(&podcastDb).Error
+	if err != nil {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return nil
+		}
+	}
+	if podcastDb.Id == "" {
+		return nil
+	}
+	return &podcastDb
+}
+
+func SavePodcast(podcast *models.Podcast) {
+	db.Create(&podcast)
+}
