+++ internal/enum/podcastType.go
@@ -6,3 +6,10 @@ const (
 	PLAYLIST PodcastType = "PLAYLIST"
 	CHANNEL  PodcastType = "CHANNEL"
 )
+
+type PodcastFetchType string
+
+const (
+	DATE    PodcastFetchType = "DATE"
+	DEFAULT PodcastFetchType = "DEFAULT"
+)
