+++ internal/models/podcast.go
@@ -3,6 +3,7 @@ package models
 import (
 	"time"
 
+	log "github.com/labstack/gommon/log"
 	"google.golang.org/api/youtube/v3"
 )
 
@@ -11,7 +12,7 @@ type PodcastEpisode struct {
 	YoutubeVideoId     string        `json:"youtube_video_id" gorm:"index:youtubevideoid_type"`
 	EpisodeName        string        `json:"episode_name"`
 	EpisodeDescription string        `json:"episode_description"`
-	PublishedDate      string        `json:"published_date"`
+	PublishedDate      time.Time     `json:"published_date"`
 	Type               string        `json:"type" gorm:"index:youtubevideoid_type_channelid_type"`
 	PodcastId          string        `json:"podcast_id" gorm:"foreignkey:PodcastId;association_foreignkey:Id"`
 	Duration           time.Duration `json:"duration"`
@@ -38,22 +39,30 @@ type EpisodePlaybackHistory struct {
 }
 
 func NewPodcastEpisodeFromPlaylist(youtubeVideo *youtube.PlaylistItem) PodcastEpisode {
+	publishedAt, err := time.Parse("2006-01-02T15:04:05Z07:00", youtubeVideo.Snippet.PublishedAt)
+	if err != nil {
+		log.Error(err)
+	}
 	return PodcastEpisode{
 		YoutubeVideoId:     youtubeVideo.Snippet.ResourceId.VideoId,
 		EpisodeName:        youtubeVideo.Snippet.Title,
 		EpisodeDescription: youtubeVideo.Snippet.Description,
-		PublishedDate:      youtubeVideo.Snippet.PublishedAt,
+		PublishedDate:      publishedAt,
 		Type:               "PLAYLIST",
 		PodcastId:          youtubeVideo.Snippet.PlaylistId,
 	}
 }
 
 func NewPodcastEpisodeFromSearch(youtubeVideo *youtube.Video, duration time.Duration) PodcastEpisode {
+	publishedAt, err := time.Parse("2006-01-02T15:04:05Z07:00", youtubeVideo.Snippet.PublishedAt)
+	if err != nil {
+		log.Error(err)
+	}
 	return PodcastEpisode{
 		YoutubeVideoId:     youtubeVideo.Id,
 		EpisodeName:        youtubeVideo.Snippet.Title,
 		EpisodeDescription: youtubeVideo.Snippet.Description,
-		PublishedDate:      youtubeVideo.Snippet.PublishedAt,
+		PublishedDate:      publishedAt,
 		Type:               "CHANNEL",
 		PodcastId:          youtubeVideo.Snippet.ChannelId,
 		Duration:           duration,
