+++ internal/models/request.go
@@ -1,0 +1,10 @@
+package models
+
+import (
+	"time"
+)
+
+type RssRequestParams struct {
+	Limit *int
+	Date  *time.Time
+}
